+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * start implementing R5RS spec
 * streams example
 * `<`, `>`, `==`, `<=` and `>=` now have variable number of arguments
+* new `throw` function that throw JS exception
 ### Bug fixes
 * fix list function [#27](https://github.com/jcubic/lips/issues/27)
 * fix `(begin (gensym))` macro that evaluate to symbol